+++ src/evm/SolidityContract.ts
@@ -59,8 +59,11 @@ export default class SolidityContract {
             }
         });
 
-        if (options.data)
-            this.options.data = options.data;
+        if (options) {
+            this.options.data = options.data || this.options.data;
+            this.options.gas = options.gas || this.options.gas;
+            this.options.gasPrice = options.gasPrice || this.options.gasPrice;
+        }
 
         if (this.options.data) {
             let encodedData: string;
@@ -72,13 +75,12 @@ export default class SolidityContract {
                 from: this.controller.defaultAddress,
                 data: encodedData
             }, this.controller)
-                .send({
-                    gas: options.gas,
-                    gasPrice: options.gasPrice
-                })
+                .gas(this.options.gas)
+                .gasPrice(this.options.gas)
+                .send()
                 .then((receipt: TXReceipt) => {
                     this.receipt = receipt;
-                    return this.address(this.receipt.contractAddress);
+                    return this.setAddressAndPopulate(this.receipt.contractAddress);
                 })
         } else {
             throw errors.InvalidDataFieldInOptions();
@@ -91,6 +93,18 @@ export default class SolidityContract {
      * @param {string} address The address to assign to the contract
      * @returns {SolidityContract} The contract
      */
+    setAddressAndPopulate(address: string): SolidityContract {
+        this.options.address = address;
+        this._attachMethodsToContract();
+        return this
+    }
+
+    /**
+     * Sets the address of the contract.
+     *
+     * @param {string} address The address to assign to the contract
+     * @returns {SolidityContract} The contract
+     */
     address(address: string): SolidityContract {
         this.options.address = address;
         this._attachMethodsToContract();
@@ -98,7 +112,7 @@ export default class SolidityContract {
     }
 
     /**
-     * Sets the gas of the contract.
+     * Sets the default gas for the contract.
      *
      * @param {number} gas The gas to assign to the contract
      * @returns {SolidityContract} The contract
@@ -109,7 +123,7 @@ export default class SolidityContract {
     }
 
     /**
-     * Sets the gas price for deploying the contract.
+     * Sets the default gas price for the contract.
      *
      * @param {number} gasPrice The gas price to assign to the contract
      * @returns {SolidityContract} The contract